Made the wire prefab. Added collider to hydro pole. Draws gizmos on wire points. Can modify the name of wire points and the radius at which they're condiered "connected" to the wire. Fixed #4 by converting the wire point to the wire's local space.
We were comparing the wire points to the wire's start and end points. If it was too far, it would break it. But the problem was that they were all too far because the wire is in local space. By using InverseTransformPoint() on the wire point (where the wire is supposed to be connected to on the pole), it turns it local, and the radius set in the inspector now more accurately represents how far the wire needs to be before being considered "broken"