Fixed #2 Collider and Layer Issue
The problem was that we were temporarity changing the layer to "Ignore Raycasts" layer when placing, then turning it back to it's original layer upon placed. When we changed to the ignore raycast layer, it was only for the root object, not it's children. If the children had the colliders, the raycast would still hit them because they were still on the non-ignore raycast layer.

Instead of playing with layers, we now just enable and disable all the colliders unpon begin placing and end placing respectively.

It should be noted this is not a perfect solution. If we're placing objects with colliders disabled by default, they will be turned on when placing is complete.